Remove circular reference from ParticipantType model
Remove the mLabelProperty reference in the ParticipantType model.  There is a bug in active android that causes a circular reference on one-to-one model ownerships:
https://github.com/pardom/ActiveAndroid/issues/107

Instead, add a boolean to the Property model (mUseAsLabel).  If this is true, the the value of this property will be used as the label for a participant.